requirements.txt
numpy
openai
scikit-learn
pyyaml
IMAPClient
beautifulsoup4
requests
backoff
